aider: Replace "Upload Documents" button with file upload option.
# Aider chat conversation:

USER: make the `Upload Documents` button open a file upload option so that the user can upload their documents

ASSISTANT: To implement a file upload feature, we need to modify the `components/header.tsx` file. We will replace the "Upload Documents" button with an input of type "file". This will allow the user to select files from their device.

Here is the edit block for the change:

```tsx
components/header.tsx
<<<<<<< HEAD
        <div
          className={cn(buttonVariants())}
        >
          <IconVercel className="mr-2" />
          <span className="hidden sm:block">Upload Documents</span>
          <span className="sm:hidden">Upload</span>
        </div>
=======
        <label
          className={cn(buttonVariants())}
        >
          <IconVercel className="mr-2" />
          <span className="hidden sm:block">Upload Documents</span>
          <span className="sm:hidden">Upload</span>
          <input type="file" className="hidden" multiple onChange={handleFileUpload} />
        </label>
>>>>>>> updated
```

In the updated code, we've replaced the div with a label and added an input of type "file". The `handleFileUpload` function will be called when the user selects files. This function needs to be implemented to handle the file upload process.

Here is the edit block for adding the `handleFileUpload` function:

```tsx
components/header.tsx
<<<<<<< HEAD
export async function Header() {
  const session = await auth()
=======
export async function Header() {
  const session = await auth()

  const handleFileUpload = (event) => {
    const files = event.target.files;
    // TODO: Implement file upload process
  }
>>>>>>> updated
```

In the updated code, we've added the `handleFileUpload` function. This function receives an event object, from which we can get the selected files. The file upload process needs to be implemented in the TODO section.