aider: Reverted changes to remove file upload functionality from the header component.
# Aider chat conversation:

USER: can you revert those last changes?

ASSISTANT: To revert the changes, we need to modify the `components/header.tsx` file. We will replace the file upload input with the "Upload Documents" button.

Here is the edit block for the change:

```tsx
components/header.tsx
<<<<<<< HEAD
        <label
          className={cn(buttonVariants())}
        >
          <IconVercel className="mr-2" />
          <span className="hidden sm:block">Upload Documents</span>
          <span className="sm:hidden">Upload</span>
          <input type="file" className="hidden" multiple onChange={handleFileUpload} />
        </label>
=======
        <div
          className={cn(buttonVariants())}
        >
          <IconVercel className="mr-2" />
          <span className="hidden sm:block">Upload Documents</span>
          <span className="sm:hidden">Upload</span>
        </div>
>>>>>>> updated
```

In the updated code, we've replaced the label with a div and removed the input of type "file".

Here is the edit block for removing the `handleFileUpload` function:

```tsx
components/header.tsx
<<<<<<< HEAD
export async function Header() {
  const session = await auth()

  const handleFileUpload = (event) => {
    const files = event.target.files;
    // TODO: Implement file upload process
  }
=======
export async function Header() {
  const session = await auth()
>>>>>>> updated
```

In the updated code, we've removed the `handleFileUpload` function.