aider: Replaced "Upload Documents" button with an input element of type "file" to handle file upload.
# Aider chat conversation:

USER: so we had to revert becaue the method broke the site. can we try another way?

ASSISTANT: Absolutely, we can try another approach to implement the file upload functionality.

One way to do this is by using the HTML `input` element with `type="file"`. This will allow the user to select files from their device storage. We can then handle the file upload in the `onChange` event of the `input` element.

Here are the steps we need to follow:

1. Replace the current "Upload Documents" button with an `input` element of `type="file"`.
2. Add an `onChange` event handler to the `input` element to handle the file upload.

Let's make these changes in the `components/header.tsx` file.

Here is the *edit block* for the changes:

```tsx
components/header.tsx
<<<<<<< HEAD
        <div
          className={cn(buttonVariants())}
        >
          <IconVercel className="mr-2" />
          <span className="hidden sm:block">Upload Documents</span>
          <span className="sm:hidden">Upload</span>
        </div>
=======
        <input
          type="file"
          className={cn(buttonVariants())}
          onChange={(event) => {
            const file = event.target.files[0];
            console.log(file);
            // TODO: Handle the file upload
          }}
        />
>>>>>>> updated
```

In the `onChange` event handler, we're currently just logging the selected file to the console. You'll need to replace this with the actual file upload logic, depending on how you want to handle the file upload.